requirements.txt
Flask
cohere
python-dotenv
gunicorn